stress-ng.h: make thermal zone type_instance a uint32_t
This helps clean up thermal zone type conversion and does not waste
any space as the uint16_t was padded for alignment.

+++ stress-ng.h
@@ -2289,7 +2289,7 @@ typedef struct {
 typedef struct stress_tz_info {
 	char	*path;			/* thermal zone path */
 	char 	*type;			/* thermal zone type */
-	uint16_t type_instance;		/* thermal zone instance # */
+	uint32_t type_instance;		/* thermal zone instance # */
 	size_t	index;			/* thermal zone # index */
 	struct stress_tz_info *next;	/* next thermal zone in list */
 } stress_tz_info_t;
